🌱completes story 4 (garden metadata creation)🌱
`poetry add pydantic`

pydantic models for (many) datacite fields

new file `metadata.py` with pydantic definitions, contents are currently
just what was generated with
https://github.com/koxudaxi/datamodel-code-generator/, which I fed the
datacite 4.3 json schema found at
https://github.com/datacite/schema/blob/da8e10655cd9d6fffadf11fee142d406fb6f59cc/source/json/kernel-4.3/datacite_4.3_schema.json.
Note that they don't have a similar json schema for 4.4 (only xsd) so
whatever we keep from this this should probably be double-checked for
compatibility. If it turns out the 4.4 xsd is easy to convert to json
schema there's no reason not to, I just didn't find an easy way to do it.

Implement basic Garden pydantic model

The new file `model.py` contains a much, much simpler pydantic model
than the auto-generated one from the datacite 4.3 schema.

It also contains a `TempField` enum, just for use with default values
like we discussed during the gathertown meeting Tuesday.

The model is certainly not complete, but the next steps that I see are
likely to be a bit more involved and/or require some hand-holding
for a design decision.

These are obviously not all the fields we'll want
eventually - I'm not even sure if these are all the fields we
want *today*. These were just the fields I was pretty sure were safe to
include, and I put a handful of non-required fields in there too so that
we had some `TempField.REQUIRED`s and some `TempField.WARN`s

In particular, the actual validation can occur at a few different places
and can be more or less involved, depending on how much work we want a
validator to do for us.

For example, I think it would be straightforward to have the
validator for the doi just automatically try and hit the datacite API if
the user didn't explicitly provide one, OR we could force the user to
call a distinct register_new_doi function (which would naturally perform
some validation when it builds the json payload for the request).

Similarly, I think we'll want the `creators` field to be more than just
plain strings and while I think the `nameparser` library looks really
promising and I'd love to build that into the validation step, I don't
know if it's worth tying ourselves to it yet.

adding comments, minimal `create_garden()` method

implements method to write metadata.json to cwd

Couple of ideas here, both plausibly Bad Ideas.

The first is a validator `check_tempfield`. This is called on *every*
explicit field update, just to check for values that are still
`TempFields`, which should only ever show up as unchecked defaults (i.e.
only as the first arg to a pydantic `Field` function invocation). The
idea is that we want a validator to catch the unset-default
values *eventually*, but not until the user has done something to
indicate that they think they've provided enough info.

If a user is invoking a `register_garden` method, they presumably think
they're done defining the model metadata, so we verify that by
1. exporting their model to a python dict
2. feeding that dict into the `Garden` constructor as **kwargs

The idea is that any still-unset defaults would look (to pydantic) like
someone manually setting e.g.  `title=TempField.REQUIRED`, which will
trigger the `check_tempfield` validator.

The exceptions that validators raise should already be pretty
logs-friendly and human-friendly, if this works how I think it does.

poetry fix

typing typo

flake8 keeping me honest

fix: doi check was skipping when given regex

I'm not sure if this is me misunderstanding that argument, or if this is
a pydantic bug, but it's not the end of the world to take it out

updated `create_garden` method

add `Garden`to `__all__` export list

refactor Garden to not use TempField defaults hack

implement `validate` method for Garden

Note that this shadows a classmethod provided by pydantic's BaseModel,
which also happens to be the only undocumented BaseModel helper
function. I'm also happy to note that the hacky-feeling
`_ = Garden(**garden_to_validate)` trick to force validation is exactly
how the library method does it under the hood.

I'd even go as far as to say that I'm feeling pretty validated, myself.

refactor/cleanup model.py and register_garden

pytest fixtures for varyingly incomplete gardens

🐛 tests were probably a good idea all along

cleanup, more tests

validate year field

real doi prefix

cleanup + test json

refactor `creator` -> `author`

fix: forgot to return validated year str

remove owen-specific deps from poetry