invoke pipeline from garden object (#125)
* move clean_identifier to utils

* pipelines have `short_name` attribute, ...

`shortname` variables elsewhere renamed to `short_name`

if one isn't set explicitly, it's inferred from the variable name used
in the python file which defines the pipeline being registered

* gardens can access their pipelines as ✨attributes✨

* garden tab-completes pipelines, like any attribute

... in IDEs that use `dir` to provide autocompletion, which is at least
the regular python repl as well as ipython/jupyter notebooks

* fix tests

* gardens now have read-only `.pipelines` and ...

`.pipeline_names` attributes. These are regular @properties, so they're ignored
by pydantic and are only used when interacting with an actual instance of
a garden. These are mostly for convenience in a notebook/repl, where
the actual short_names of pipelines also work as read-only attributes.

* add --alias option to add-pipeline command

* new tests

* respond to PR feedback