Pin commit hash in model connectors
Both `HFConnector` and `GitHubConnector` now attempt to retreive the commit hash and store it in their ModelMetadata.model_version when instantiated. If no commmit hash is found, the model.revision and ModelMetadata.model_version fields are set to their default value.

The pinned commit hash is then used when `stage()` is called to pull down the same model each time the entrypoint is run.

New unit tests in `tests/test_models.py`